moved pandas grouping to another file from views
